backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.23
botocore==1.37.23
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.69.0
pandas==2.2.3
pydantic==2.11.1
pydantic_core==2.33.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
s3transfer==0.11.4
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
